refactor(desktop): extract turn reliability boundaries
